fix: fix vue config
